Do some int->bool conversions in mythframe.h.
Convert some variable from "int" to "bool", and convert some helper
functions to return "bool" instead of "int".  Propagate the changes to
all affected files.  These changes originated in looking for variables
that were implicitly cast to/from a boolean.

+++ mythtv/libs/libmythtv/mythccextractorplayer.cpp
@@ -88,7 +88,7 @@ void MythCCExtractorPlayer::OnGotNewFrame(void)
         double fps = frame->frame_rate;
         if (fps <= 0)
             fps = GetDecoder()->GetFPS();
-        double duration = 1 / fps + frame->repeat_pict * 0.5 / fps;
+        double duration = 1 / fps + static_cast<double>(frame->repeat_pict) * 0.5 / fps;
         m_curTime += duration * 1000;
         m_videoOutput->DoneDisplayingFrame(frame);
     }
